Various fixes for documentation
